Fixed hard reload issue.
We had previously been experiencing an issue where hard reloads caused an error in the useEffect hook of pages that required a db request. This error was due to the next/router instance not being instantiated by the time it was referenced. We can avoid this issue by checking if what we need from the router exists before making the request, and adding the router as a dependency to the useEffect. This way we perform no action before the router has loaded. We then detect when the router has loaded as it is a dependency of the useEffect and finally perform the request that relies on the router.