feat(www:Edit): validate participant emails
We weren't really doing any kind of validation of participant emails. This PR adds a validate function as a prop to the ChipForm as well as a validateMsg. When the validate function returns false for an itemKey, it will display the error message on the input. It is a bit nicer then the big bottom error banner because it draws the user to the problem in the form