fix(www): fix selected topic reactivity
When a topic is opened or closed by another user, the selected topic
would not update with the change because the entire object was part of
the page's state. This can be avoided by only using the id in the
state and finding the selected topic each time the component updates.